clip_service/requirements.txt
fastapi
uvicorn
torch
open_clip_torch
pillow
